+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.0.1 (02.10.2019)
+
+* Added resolving of Addon dependencies of other Addons
+* Removed some ugly hacks around ZombieBox file watching.
+
 ## 1.0.0 (09.08.2019)
 
 Initial release